requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
waitress>=2.1.0
requests>=2.31.0
python-decouple>=3.8
django-extensions>=3.2.0
whitenoise>=6.5.0